requirements.txt
pandas==1.4.1
requests==2.27.1
transformers==4.16.2
sentence-transformers==2.2.0
pydantic==1.9.0
fastapi>=0.75.0
uvicorn>=0.17.6
streamlit>=1.9.0
faiss-cpu>=1.7.2
numpy>=1.22.3